+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes in version 0.9 (dev)
 
 * Avoid ``UrlNode.DoesNotExist`` exception when using ``{% render_breadcrumb %}`` on 404 pages.
 * Change ``slug`` size to 100 characters.
+* Added ``FluentContentsPage.create_placeholder(slot)`` API.
 
 
 Released in 0.9b4: